refactor: 거리타입 Double형으로 변경

+++ hub-service/src/main/java/com/gbg/hubservice/domain/entity/HubRoute.java
@@ -13,6 +13,6 @@ public class HubRoute extends BaseEntity {
     private UUID id;
     private UUID startHubId;
     private UUID endHubId;
-    private Integer distance;
+    private double distance;
 
 }
